Slack is out, Gitter is in
Slack and Gitter are similar services (chat with some extra
integrations and whatnot). The main difference is that Slack
is by default for closed teams. And trying to make it open is
a pain (that 'request a Slack invite thing is extra friction).

Gitter is open by default. You can just go to
[the freesewing community on
Gitter](https://gitter.im/freesewing/freesewing) and
read along. You don't even have to log in, unless you want
to post, that is :)

So, today we talked a bit about it and we (that would be @cabi
and myself decided to ditch Slack in favour of Gitter.

This commit does updates the website so that instead of Slack
it now references Gitter everywhere.