Changed behaviour of user signup and welcome page
When people confirm their account, they are now automatically logged in.

In addition, they are still taken to the welcome page as before, but
the welcome page now loads a modal window that allows you to quickly
configure your units, username, and profile picture.

Once that's done, it takes you to your account page.