A bunch of changes in preparation of the new year rollout of patrons.
There's also a bunch of small changes to fix issues reported by the
htmlproofer.

Some stuff has also gone, like the donate page (redirects to the patron sign up page).

I've also removed the list of contributors, because I can't maintain it
manually. Contributions are listed on GitHub, so there's that.

I've also updated the donations pledge page with details on the why/how
of money.