Update app.yaml
I`m not sure if you`d like to keep the word "patron" as it is in english, because it is a "common expression". In the german-speaking Countries there are only the "very" young people who understand what a patron is. A "Gönner" on the other hand is a very polite an positive german word for the same thing.